requirements.txt
python-logging-loki>=0.3.1
langchain>=0.0.329
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-langchain>=0.0.1b
opentelemetry-exporter-otlp>=1.20.0
opentelemetry-exporter-prometheus>=1.20.0
sqlalchemy>=2.0.23
requests>=2.31.0
openai>=1.6.1
python-dotenv>=1.0.0
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.2
